fastapi_project/requirements.txt

    fastapi
    uvicorn
    phidata
    google-search-results
    google-generativeai
    openai
    huggingface_hub